Move CoreData to core package
